+++ src/components/Button.tsx
@@ -1,38 +1,32 @@
-import styled, {css} from "styled-components";
+import React from 'react';
+import classNames from 'classnames';
 import theme from '../theme';
-import { applyButtonVariantStyles, ButtonVariant } from "../theme/buttons";
+import { getButtonVariantStyles, ButtonVariant } from "../theme/buttons";
+import './Button.css';
 
-export { applyButtonVariantStyles };
-export const buttonCss = css<{variant?: ButtonVariant}>`
-  ${props => applyButtonVariantStyles(props.variant || 'primary')}
+// Re-export for backwards compatibility
+export { applyButtonVariantStyles } from "../theme/buttons";
 
-  font-size: 1.6rem;
-  line-height: 2rem;
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
-  display: inline-flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  height: 4rem;
-  padding: 0 3rem;
-  border: 0;
-  border-radius: 0.5rem;
-  box-shadow: 0px 0.2rem 0.4rem rgba(0, 0, 0, 0.2);
-  transition: all 0.2s ease-in-out;
+// Export buttonCss as a string for backwards compatibility
+export const buttonCss = 'button-base';
+
+/**
+ * Link style CSS fragment for backwards compatibility.
+ *
+ * @deprecated This export is deprecated. Use the ButtonLink component instead.
+ * @note Breaking change: Previously returned a styled-components css fragment,
+ * now returns a plain CSS string. This is part of the migration away from styled-components.
+ * The string can still be used in styled-components template literals.
+ */
+export const linkStyle = `
+  color: ${theme.colors.link.color};
+  cursor: pointer;
   text-decoration: none;
-  user-select: none;
-  white-space: nowrap;
 
-  &:not([disabled]) {
-    cursor: pointer;
-  }
-  &:disabled {
-    opacity: 0.4;
-  }
-
-  & + & {
-    margin-left: 1.6rem;
+  :hover,
+  :focus {
+    text-decoration: underline;
+    color: ${theme.colors.link.hover};
   }
 `;
 
@@ -53,53 +47,116 @@ export interface WaitingButtonProps extends ButtonBase {
   waitingText: string;
 }
 
-export const Button = styled((props: ButtonProps | WaitingButtonProps) => {
-  const {
-    disabled,
-    isWaiting,
-    waitingText,
-    children,
-    variant,
-    ...otherProps
-  } = props;
+export const Button = React.forwardRef<HTMLButtonElement, ButtonProps | WaitingButtonProps>(
+  (props, ref) => {
+    const {
+      disabled,
+      isWaiting,
+      waitingText,
+      children,
+      variant = 'primary',
+      className,
+      style,
+      ...otherProps
+    } = props;
 
-  return <button
-    {...otherProps}
-    disabled={isWaiting || disabled}
-  >
-    {(isWaiting && waitingText) || children}
-  </button>;
-})`
-  ${buttonCss}
-`;
+    const variantStyles = getButtonVariantStyles(variant);
+    const buttonStyle = {
+      '--button-bg': variantStyles.background,
+      '--button-bg-hover': variantStyles.backgroundHover,
+      '--button-bg-active': variantStyles.backgroundActive,
+      '--button-color': variantStyles.color,
+      '--button-outline': variantStyles.outline,
+      '--button-shadow': variantStyles.shadow,
+      '--button-font-weight': variantStyles.fontWeight ?? 700,
+      ...style
+    } as React.CSSProperties;
 
-export const LinkButton = styled(({ variant, ...props}: LinkButtonBase) =>
-  <a {...props}>{props.children}</a>
-)`
-  ${buttonCss}
-`;
-
-export const linkStyle = css`
-  color: ${theme.colors.link.color};
-  cursor: pointer;
-  text-decoration: none;
-
-  :hover,
-  :focus {
-    text-decoration: underline;
-    color: ${theme.colors.link.hover};
+    return (
+      <button
+        {...otherProps}
+        ref={ref}
+        disabled={isWaiting || disabled}
+        className={classNames('button-base', className)}
+        style={buttonStyle}
+      >
+        {(isWaiting && waitingText) || children}
+      </button>
+    );
   }
-`;
+);
 
-export const PlainButton = styled.button`
-  cursor: pointer;
-  border: none;
-  margin: 0;
-  padding: 0;
-  background: none;
-`;
+Button.displayName = 'Button';
 
-// tslint:disable-next-line:variable-name
-export const ButtonLink = styled(PlainButton)`
-  ${linkStyle}
-`;
+export const LinkButton = React.forwardRef<HTMLAnchorElement, LinkButtonBase>(
+  (props, ref) => {
+    const { variant = 'primary', className, style, children, ...otherProps } = props;
+
+    const variantStyles = getButtonVariantStyles(variant);
+    const linkStyle = {
+      '--button-bg': variantStyles.background,
+      '--button-bg-hover': variantStyles.backgroundHover,
+      '--button-bg-active': variantStyles.backgroundActive,
+      '--button-color': variantStyles.color,
+      '--button-outline': variantStyles.outline,
+      '--button-shadow': variantStyles.shadow,
+      '--button-font-weight': variantStyles.fontWeight ?? 700,
+      ...style
+    } as React.CSSProperties;
+
+    return (
+      <a
+        {...otherProps}
+        ref={ref}
+        className={classNames('button-base', className)}
+        style={linkStyle}
+      >
+        {children}
+      </a>
+    );
+  }
+);
+
+LinkButton.displayName = 'LinkButton';
+
+export const PlainButton = React.forwardRef<HTMLButtonElement, React.ComponentPropsWithoutRef<'button'>>(
+  (props, ref) => {
+    const { className, children, ...otherProps } = props;
+    return (
+      <button
+        {...otherProps}
+        ref={ref}
+        className={classNames('plain-button', className)}
+      >
+        {children}
+      </button>
+    );
+  }
+);
+
+PlainButton.displayName = 'PlainButton';
+
+export const ButtonLink = React.forwardRef<HTMLButtonElement, React.ComponentPropsWithoutRef<'button'>>(
+  (props, ref) => {
+    const { className, style, children, ...otherProps } = props;
+
+    const linkStyleVars = {
+      '--link-color': theme.colors.link.color,
+      '--link-hover-color': theme.colors.link.hover,
+      ...style
+    } as React.CSSProperties;
+
+    return (
+      <button
+        {...otherProps}
+        ref={ref}
+        className={classNames('button-link', className)}
+        style={linkStyleVars}
+      >
+        {children}
+      </button>
+    );
+  }
+);
+
+ButtonLink.displayName = 'ButtonLink';